Add TAS logic

+++ loadtas.h
@@ -2,6 +2,7 @@
 
 #include <vector>
 #include <string>
+#include <cstdint>
 
-extern std::vector<int> tas_inputs;
+extern std::vector<uint8_t> tas_inputs;
 extern void load_tas(std::string filename);
